// src/app/models/user.model.ts
export class User {
  id: number;
  name: string;
  apellido: string;
  dni: string;
  email: string;
  genero: string;
  fecha_nacimiento: Date;
  password: string;
  password_confirmation: string;
  provincia: string;
  localidad_id: string;
  municipio_id: string;
  ether_address: string;
  rol_id: string;
  dni_image: any;
  user_verified: number;
}